Renames the bigip_configsync_actions module
Retains the old name by making a symlink. We can remove it in a later
release.
