+++ test/integration/targets/bigip_configsync_action/defaults/main.yaml
@@ -1,0 +1,3 @@
+---
+
+device_group: "sdbt_sync_failover_dev_group"
