+++ test/integration/targets/bigip_configsync_action/tasks/test-device-to-group.yaml
@@ -1,0 +1,23 @@
+---
+
+- name: Sync configuration from device to group
+  bigip_configsync_actions:
+      device_group: "{{ device_group }}"
+      sync_device_to_group: yes
+  register: result
+
+- name: Sync configuration from device to group
+  assert:
+      that:
+          - result|changed
+
+- name: Sync configuration from device to group - Idempotent check
+  bigip_configsync_actions:
+      device_group: "{{ device_group }}"
+      sync_device_to_group: yes
+  register: result
+
+- name: Sync configuration from device to group - Idempotent check
+  assert:
+      that:
+          - not result|changed
